chore: cleanup `.prettierignore` file
A few invalid entries have been removed from the `.prettierignore`
file, and the remaining entries have been alphabetized so that they are
easier to quickly scan.

Additionally, a leading slash has been added for any entries that are
intended to be relative to the repository root directory. This was so
that it would be more explicit which patterns would apply to
subdirectories (this ambiguity led to me missing the
`test/test-results` directory on the first pass).

I did this because I was investigating a strange ESLint error, and
these entries made the investigation a little more complicated than it
needed to be.