test: Fix test config
Our Mocha and Jest configuration has been messy because we're in the
middle of a migration from Mocha to Jest. Each file had to be
explicitly included in the Jest configuration or ignored in the Mocha
configuration. This was inconvenient and error-prone, and resulted in
some tests not being run at all.

Both test configurations have been updated to use a shared list of
Mocha test files. There are only a few of these files left, and this
list should only get shorter as we migrate more tests to Jest. No
further configuration changes will be needed to add Jest tests.